Organize root feature using atomic design
