+++ app/[locale]/page.tsx
@@ -1,4 +1,4 @@
-import RootPage from '@/components/ui/root/RootPage';
+import RootPage from '@/components/ui/root/pages/RootPage';
 
 export default async function Home() {
   return (
